backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
supabase==2.11.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.1
openai==1.63.2
httpx==0.28.1
